requirements.txt
Django>=4
djangorestframework
psycopg2-binary
celery
django-cors-headers
pandas
openpyxl
gunicorn
